+++ src/collective/recipe/backup/tests/test_docs.py
@@ -68,6 +68,7 @@ def test_suite():
         'multiple.rst',
         'no_rsync.rst',
         'options.rst',
+        'zipbackup.rst',
         ]
     for docfile in docfiles:
         suite.addTest(doctest.DocFileSuite(